+++ src/common/engine.js
@@ -175,7 +175,10 @@ export class CurrencyEngine {
       this.log(
         'Could not load transactionList file. Failure is ok on new device'
       )
-      await disklet.setText(TXID_MAP_FILE, JSON.stringify(this.txIdMap))
+      await disklet.setText(
+        TRANSACTION_STORE_FILE,
+        JSON.stringify(this.transactionList)
+      )
     }
 
     let isEmptyTransactions = true
@@ -197,6 +200,11 @@ export class CurrencyEngine {
           this.addTransaction(cc, edgeTransaction)
         }
       }
+    }
+    for (const currencyCode in this.transactionList) {
+      this.walletLocalData.numTransactions[currencyCode] = this.transactionList[
+        currencyCode
+      ].length
     }
   }
 
@@ -273,7 +281,7 @@ export class CurrencyEngine {
     edgeTransaction: EdgeTransaction,
     lastSeenTime?: number
   ) {
-    this.log('executing addTransaction: ', edgeTransaction)
+    this.log('executing addTransaction: ', edgeTransaction.txid)
     // remove SPAM and proxy allowance transactions (ie DEX extra transaction)
     // this should reduce confusion for users
     if (
@@ -322,6 +330,10 @@ export class CurrencyEngine {
       }
       // add transaction to list of tx's, and array of changed transactions
       this.transactionList[currencyCode].push(edgeTransaction)
+      this.walletLocalData.numTransactions[currencyCode] = this.transactionList[
+        currencyCode
+      ].length
+      this.walletLocalDataDirty = true
 
       this.transactionListDirty = true
       this.transactionsChangedArray.push(edgeTransaction)
@@ -751,10 +763,10 @@ export class CurrencyEngine {
     const cleanOptions = asCurrencyCodeOptions(options)
     const { currencyCode = this.currencyInfo.currencyCode } = cleanOptions
 
-    if (this.transactionList[currencyCode] == null) {
+    if (this.walletLocalData.numTransactions[currencyCode] == null) {
       return 0
     } else {
-      return this.transactionList[currencyCode].length
+      return this.walletLocalData.numTransactions[currencyCode]
     }
   }