+++ src/fio/fioInfo.js
@@ -7,25 +7,26 @@ import { imageServerUrl } from '../common/utils'
 
 const defaultSettings: any = {
   apiUrls: [
-    'https://fio.greymass.com:443/v1/',
-    'https://fio.zenblocks.io:443/v1/',
-    'https://api.fio.alohaeos.com:443/v1/',
-    'https://fio-mainnet.eosblocksmith.io:443/v1/',
-    'https://fio.eu.eosamsterdam.net:443/v1/',
-    'https://fio.eosdac.io:443/v1/',
+    'https://fio.eu.eosamsterdam.net/v1/',
+    'https://fio.eosdac.io/v1/',
     'http://fioapi.nodeone.io:6881/v1/',
-    'https://fio.acherontrading.com:443/v1/',
-    'https://fio.eos.barcelona:443/v1/',
-    'https://fio.eosusa.news:443/v1/',
-    'https://fio.eosargentina.io:443/v1/',
-    'https://api.fio.currencyhub.io:443/v1/',
-    'https://fio.eoscannon.io:443/v1/',
-    'https://fio.eossweden.org:443/v1/',
-    'https://fio.maltablock.org:443/v1/',
-    'https://api.fio.eosdetroit.io:443/v1/',
-    'https://fio.eosdublin.io:443/v1/',
-    'https://fio.eosphere.io:443/v1/',
-    'https://fio.cryptolions.io:443/v1/'
+    'https://fio.eosphere.io/v1/',
+    'https://fio.eosrio.io/v1/',
+    'https://fio.acherontrading.com/v1/',
+    'https://fio.eos.barcelona/v1/',
+    'https://api.fio.eosdetroit.io/v1/',
+    'https://fio.zenblocks.io/v1/',
+    'https://api.fio.alohaeos.com/v1/',
+    'https://fio.greymass.com/v1/',
+    'https://fio.eosusa.news/v1/',
+    'https://fio.eosargentina.io/v1/',
+    'https://fio.cryptolions.io/v1/',
+    'https://fio-mainnet.eosblocksmith.io/v1/',
+    'https://api.fio.currencyhub.io/v1/',
+    'https://fio.eoscannon.io/v1/',
+    'https://fio.eosdublin.io/v1/',
+    'https://api.fiosweden.org/v1/',
+    'https://fio.maltablock.org/v1/'
   ],
   historyNodeUrls: [
     'https://fio.greymass.com/v1/',
@@ -36,8 +37,15 @@ const defaultSettings: any = {
   ],
   fioRegApiUrl: 'https://reg.fioprotocol.io/public-api/',
   fioDomainRegUrl: 'https://reg.fioprotocol.io/domain/',
-  defaultRef: 'edgefree',
-  fallbackRef: 'edge'
+  fioAddressRegUrl: 'https://reg.fioprotocol.io/address/',
+  defaultRef: 'edge',
+  fallbackRef: 'edge',
+  freeAddressRef: 'edgefree',
+  errorCodes: {
+    INVALID_FIO_ADDRESS: 'INVALID_FIO_ADDRESS',
+    FIO_ADDRESS_IS_NOT_EXIST: 'FIO_ADDRESS_IS_NOT_EXIST',
+    FIO_ADDRESS_IS_NOT_LINKED: 'FIO_ADDRESS_IS_NOT_LINKED'
+  }
 }
 
 export const currencyInfo: EdgeCurrencyInfo = {
@@ -49,8 +57,8 @@ export const currencyInfo: EdgeCurrencyInfo = {
 
   defaultSettings,
 
-  addressExplorer: 'https://explorer.fioprotocol.io/pubkey/%s',
-  transactionExplorer: 'https://explorer.fioprotocol.io/transaction/%s',
+  addressExplorer: 'https://fio.bloks.io/key/%s',
+  transactionExplorer: 'https://fio.bloks.io/transaction/%s',
 
   denominations: [
     // An array of Objects of the possible denominations for this currency