+++ test/engine/engine.test.js
@@ -65,6 +65,9 @@ for (const fixture of fixtures) {
     onTransactionsChanged(transactionList) {
       // console.log('onTransactionsChanged:', transactionList)
       emitter.emit('onTransactionsChanged', transactionList)
+    },
+    onAddressChanged() {
+      emitter.emit('addressChanged')
     }
   }
 
@@ -144,7 +147,7 @@ for (const fixture of fixtures) {
 
   describe('Start engine', function () {
     it('Get BlockHeight', function (done) {
-      this.timeout(10000)
+      this.timeout(100000)
       emitter.once('onBlockHeightChange', height => {
         const thirdPartyHeight = 1578127
         // this validation is not OK for RSK
@@ -204,6 +207,10 @@ const callbacks: EdgeCurrencyEngineCallbacks = {
   onTransactionsChanged(transactionList) {
     // console.log('onTransactionsChanged:', transactionList)
     emitter.emit('onTransactionsChanged', transactionList)
+  },
+  onAddressChanged() {
+    // console.log('onTransactionsChanged:', transactionList)
+    emitter.emit('addressChanged')
   }
 }